further abstracted calculating arrow end points from using na and frame transforms. not in plot anymore
