+++ src/utils/plot.rs
@@ -1,11 +1,9 @@
 #[allow(unused)]
 // rust robotics
 use crate::utils::defs;
-use crate::utils::transforms::FrameTransform3;
 
 // 3rd party or std
 use minifb;
-use nalgebra as na;
 use plotters::prelude::*;
 use plotters::{chart::ChartState, coord::types::RangedCoordf64};
 use plotters_arrows;
@@ -156,23 +154,14 @@ pub fn rectangle_filled_element(
 }
 #[allow(unused)]
 pub fn arrow_element(
-    start_point: &(f64, f64),
-    length: f64,
-    heading_angle: f64,
-    angle_units: defs::AngleUnits,
+    end_points: &[(f64, f64); 2],
     color: &(u8, u8, u8),
 ) -> plotters_arrows::ThinArrow<(f64, f64), i32> {
-    let tf = FrameTransform3::new(
-        &[start_point.0, start_point.1, 0., 0., 0., heading_angle],
-        angle_units,
-    );
-    let end_point = tf.point_b_to_i(&na::Point3::new(length, 0., 0.));
-
     let arrow_color = &plotters::style::RGBColor(color.0, color.1, color.2);
 
     plotters_arrows::ThinArrow::new(
-        (start_point.0, start_point.1),
-        (end_point[0], end_point[1]),
+        (end_points[0].0, end_points[0].1),
+        (end_points[1].0, end_points[1].1),
         &arrow_color,
     )
 }
